Add annual payment plan
