+++ frontend/web/components/ProjectsPage.tsx
@@ -2,6 +2,8 @@ import React, { FC } from 'react'
 import ProjectManageWidget from './ProjectManageWidget'
 import OrganisationProvider from 'common/providers/OrganisationProvider'
 import ConfigProvider from 'common/providers/ConfigProvider'
+import makeAsyncScriptLoader from 'react-async-script'
+import { onPaymentLoad } from './modals/Payment'
 
 type ProjectsPageType = {
   match: {
@@ -24,4 +26,13 @@ const ProjectsPage: FC<ProjectsPageType> = ({ match }) => {
   )
 }
 
-export default ConfigProvider(ProjectsPage)
+const InnerComponent = ConfigProvider(ProjectsPage)
+const WrappedPayment = makeAsyncScriptLoader(
+  'https://js.chargebee.com/v2/chargebee.js',
+  {
+    removeOnUnmount: true,
+  },
+)(InnerComponent)
+export default (props) => (
+  <WrappedPayment {...props} asyncScriptOnLoad={onPaymentLoad} />
+)
